Factor my submissions table into component
